+++ apps/newsletters/views.py
@@ -4,31 +4,61 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import render_to_response
 from django.template import RequestContext
 from django.core.urlresolvers import reverse
-from django.views.generic import TemplateView
-from datetime import datetime
+from django.views.generic import TemplateView, DetailView, ListView
+from django.views.generic.dates import YearArchiveView, MonthArchiveView
 
 import mailchimp
-from apps.newsletters.forms import NewsletterSignupForm
-from apps.newsletters.models import Edition
+from .forms import NewsletterSignupForm
+from .models import Edition
+
+
+class EditionYearArchiveView(YearArchiveView):
+    queryset = Edition.published_objects.all()
+    date_field = 'publish_on'
+    make_object_list = True
+
+
+class EditionMonthArchiveView(MonthArchiveView):
+    queryset = Edition.published_objects.all()
+    date_field = 'publish_on'
+    make_object_list = True
+
+
+class EditionDetailView(DetailView):
+    model = Edition
+    template_name = 'newsletters/newsletter.html'
+    slug_field = 'id'
+
+
+class EditionListView(ListView):
+    queryset = Edition.published_objects.all()
+    template_name = 'newsletters/newsletter-archive-list.html'
+
 
 class NewsletterView(TemplateView):
-    template_name='newsletters/newsletter.html'
+    template_name = 'newsletters/newsletter.html'
 
     def get_context_data(self, **kwargs):
         context = super(NewsletterView, self).get_context_data(**kwargs)
-        today=datetime.today()
         try:
             context['object'] = Edition.published_objects.latest()
         except:
             context['object'] = None
         return context
 
-letter = mailchimp.utils.get_connection().get_list_by_id(getattr(settings, 'MAILCHIMP_NEWSLETTER_LIST_ID', None))
-parents = mailchimp.utils.get_connection().get_list_by_id(getattr(settings, 'MAILCHIMP_PARENTS_LIST_ID', None))
+
+letter = mailchimp.utils.get_connection().get_list_by_id(getattr(
+    settings, 'MAILCHIMP_NEWSLETTER_LIST_ID', None))
+
+parents = mailchimp.utils.get_connection().get_list_by_id(getattr(
+    settings, 'MAILCHIMP_PARENTS_LIST_ID', None))
+
 saved_lists = {'parents': parents,
                'letter': letter}
 
-def update_subscription(list1, email_address, fname=None, lname=None, subscribe=False, unsubscribe=False):
+
+def update_subscription(list1, email_address, fname=None, lname=None,
+                        subscribe=False, unsubscribe=False):
     if email_address:
         try:
             subscriber = list1.get_member(email_address)