+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 CHANGELOG
 =========
 
+v1.2.0
+------
+
+  * Adding a dashboard for student adding of content
+
+v1.1.2
+------
+
+  * Minor fix to the newsletter archive template images
+
+v1.1.1
+------
+
+  * Adding an archive page for newsletters
+
 v1.1.0
 ------